requirements.txt
solana>=0.30.2
solders>=0.18.0
aiohttp>=3.9.0
websockets>=11.0.3
python-dotenv>=1.0.0
base58>=2.1.1
python-telegram-bot>=20.0
aiodns>=3.1.1
